Separate Vue application into components
 - Create ShortForm and Shorts components
